Ops as app
