+++ kraken/urls.py
@@ -11,9 +11,9 @@ base_urlpatterns = patterns(
     url(r'^accounts/login/$', login),
     url(r'^accounts/logout/$', logout),
     url(r'^$', 'status.views.home', name='home'),
-    url(r'^ops/$', 'status.views.ops', name='ops'),
+    url(r'^ops/$', 'ops.views.ops', name='ops'),
     url(r'^osd/(\d+)/$', 'status.views.osd_details', name='osd_details'),
-    url(r'^user/(.+?)(/.+?)?(/.+?)?/$', 'status.views.user_custom',
+    url(r'^user/(.+?)(/.+?)?(/.+?)?/$', 'ops.views.user_custom',
         name='user_custom'),
     url(r'^api/clusters/health/$', views.health, name="cluster-health"),
     url(r'^api/clusters/status/$', views.status, name="cluster-status"),
