+++ ops/templates/ops.html
